Dev to main 코드프리징 전 푸쉬

+++ app/src/main/java/com/example/accountbook/domain/model/HistoriesTotalData.kt
@@ -2,7 +2,7 @@ package com.example.accountbook.domain.model
 
 
 data class HistoriesTotalData(
-    val totalIncome: Int,
-    val totalExpense: Int,
+    val totalIncome: Long,
+    val totalExpense: Long,
     val historyList: List<HistoriesListItem>
 )
